FEM: unit tests, move mesh tests in own test module
