+++ src/Mod/Fem/CMakeLists.txt
@@ -117,6 +117,7 @@ SET(FemTests_SRCS
     femtest/__init__.py
     femtest/testccxtools.py
     femtest/testcommon.py
+    femtest/testmesh.py
     femtest/testsolverframework.py
     femtest/testtools.py
 )
